Conflicts:
	mmo/Assets/Photon Unity Networking/Demos/DemoPickup/Assets/quad-mesh.obj.meta
	mmo/Assets/Resources/Enemy/Cockatrice/kokatorisu.obj.meta
	mmo/Library/CurrentLayout.dwlt
	mmo/Library/ScriptAssemblies/Assembly-CSharp-Editor.dll.mdb
	mmo/Library/ScriptAssemblies/Assembly-CSharp-firstpass.dll.mdb
	mmo/Library/ScriptAssemblies/Assembly-CSharp.dll.mdb
	mmo/Library/assetDatabase3
	mmo/Library/expandedItems
	mmo/Library/metadata/00/00000000000000003000000000000000
	mmo/Library/metadata/00/00000000000000004000000000000000
	mmo/mmo-csharp.sln
	mmo/mmo.sln

+++ mmo/Assets/Script/Game/UI/Status/UpdateShowStatus.cs
@@ -87,7 +87,7 @@ public class UpdateShowStatus : MonoBehaviour {
         if (playerChar.GetPlayerData().HP != prevHP)
         {
             // サイズを計算する
-            float size = playerChar.GetPlayerData().HP / playerChar.GetPlayerData().MaxHP;
+            float size = (float)playerChar.GetPlayerData().HP / (float)playerChar.GetPlayerData().MaxHP;
             // サイズを変更する
             hpBarImage.transform.localScale = new Vector3(size, 1f, 1f);
             // テキストを変更する
@@ -106,7 +106,7 @@ public class UpdateShowStatus : MonoBehaviour {
         if (playerChar.GetPlayerData().SP != prevSP)
         {
             // サイズを計算する
-            float size = playerChar.GetPlayerData().SP / playerChar.GetPlayerData().MaxSP;
+            float size = (float)playerChar.GetPlayerData().SP / (float)playerChar.GetPlayerData().MaxSP;
             // サイズを変更する
             spBarImage.transform.localScale = new Vector3(size, 1f, 1f);
             // テキストを変更する
